moved serializer shortcut functions into shortcuts.py modules
