+++ intake/serializers/__init__.py
@@ -1,8 +1,7 @@
-from .request_serializer import RequestSerializer, mixpanel_request_data
-from .view_serializer import ViewMixpanelSerializer, mixpanel_view_data
+from .request_serializer import RequestSerializer
+from .view_serializer import ViewMixpanelSerializer
 from .form_submission_serializer import FormSubmissionFollowupListSerializer
-from .applicant_serializer import (
-    ApplicantMixpanelSerializer, mixpanel_applicant_data)
+from .applicant_serializer import ApplicantMixpanelSerializer
 from .organization_serializer import OrganizationSerializer
 from .app_index_serializers import (
     ApplicationIndexSerializer,
@@ -11,15 +10,14 @@ from .application_serializers import ApplicationAutocompleteSerializer
 from .status_update_serializer import StatusUpdateSerializer
 from .note_serializer import ApplicationNoteSerializer
 from .tag_serializer import TagSerializer
+from .shortcuts import (
+    mixpanel_applicant_data, mixpanel_request_data, mixpanel_view_data)
 
 
 __all__ = [
     RequestSerializer,
     ViewMixpanelSerializer,
-    mixpanel_view_data,
-    mixpanel_request_data,
     ApplicantMixpanelSerializer,
-    mixpanel_applicant_data,
     FormSubmissionFollowupListSerializer,
     OrganizationSerializer,
     ApplicationAutocompleteSerializer,
@@ -27,5 +25,8 @@ __all__ = [
     ApplicationIndexWithTransfersSerializer,
     ApplicationNoteSerializer,
     StatusUpdateSerializer,
-    TagSerializer
+    TagSerializer,
+    mixpanel_view_data,
+    mixpanel_applicant_data,
+    mixpanel_request_data
 ]
