+++ user_accounts/serializers/__init__.py
@@ -1,4 +1,5 @@
-from .user_serializer import UserMixpanelSerializer, mixpanel_user_data
+from .user_serializer import UserMixpanelSerializer
+from .shortcuts import mixpanel_user_data
 
 __all__ = [
     UserMixpanelSerializer,
