+++ user_accounts/serializers/shortcuts.py
@@ -1,0 +1,5 @@
+from .user_serializer import UserMixpanelSerializer
+
+
+def mixpanel_user_data(user):
+    return UserMixpanelSerializer(user).data
