+++ user_accounts/serializers/user_serializer.py
@@ -8,7 +8,3 @@ class UserMixpanelSerializer(serializers.Serializer):
     user_organization_name = ChainableAttributeField(
         'profile.organization.name')
     user_username = ChainableAttributeField('username')
-
-
-def mixpanel_user_data(user):
-    return UserMixpanelSerializer(user).data
